requirements.txt
psycopg2-binary==2.9.9
pandas==2.2.1
sqlalchemy==2.0.27 